Refactor ProjectDetailPage and deprecate SuiteResultsPage
- Updated ProjectDetailPage to change the back link text to "← Projects" and added a configuration label.
- Changed the "View Results" link to use a secondary button style.
- Deprecated SuiteResultsPage, replacing its content with a redirect to the dashboard.
- Removed all previous logic related to fetching and displaying suite results, including loading states and error handling.
- Added a simple redirecting message for users navigating to the deprecated page.

+++ backend/app/schemas.py
@@ -237,7 +237,7 @@ class BreakingPoints(BaseModel):
 
 
 class Headroom(BaseModel):
-    percentage: int
+    percentage: float
     explanation: str
     scale_recommendation: str
